requirements.txt
numpy==1.24.3
deeplake==3.6.6
torch>1
timm==0.9.2
pytorch-argus==1.0.0
